Testing and updates
decrypt.go - Create a goroutine specifically to monitor running processes and see if the specified executable has finished running. If so, delete the executable to prevent running it multiple times.

findprocess.go - File that uses the go-ps utility to monitor the process table. Added a function that runs through the processes and checks if the argument executable exists in it or not.

Also added a couple of test files to get practice on the golang testing framework